requirements.txt
faiss==1.8.0
numpy==2.1.1
Pillow==10.4.0
scikit_learn==1.5.2
skimage==0.0
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
